Don't call reset_session when session times out
Resetting the session will cause CSRF tokens to be out of sync when the user signs out, resulting in an InvalidAuthenticityToken exception to be raised when they sign back in.

Instead, we should call Devise's `reset_authentication_token!` at the time the session times out (in our custom timeout method), which is how Devise implements their timeoutable feature.